% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chopin.magnification-package.R
\docType{package}
\name{chopin.magnification-package}
\alias{chopin.magnification}
\alias{chopin.magnification-package}
\title{chopin.magnification: R Code for Field Methodological Comparison of Biomagnification Assessment}
\description{
Use Rmarkdown First method to build your package. Start your package with documentation. Everything can be set from a Rmarkdown file in your project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FlorenceMounier/chopin.magnification}
  \item Report bugs at \url{https://github.com/FlorenceMounier/chopin.magnification/issues}
}

}
\author{
\strong{Maintainer}: Florence Mounier \email{florence.mounier@inrae.fr} (\href{https://orcid.org/0000-0001-8916-7007}{ORCID})

}
\keyword{internal}
