% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reorder_pcb.R
\name{fct_reorder_PCB}
\alias{fct_reorder_PCB}
\title{fct_reorder_PCB Reorder PCB factor}
\usage{
fct_reorder_PCB(data)
}
\arguments{
\item{data}{dataset with column PCB}
}
\value{
data with reordered factor column
}
\description{
fct_reorder_PCB Reorder PCB factor
}
\examples{
levels(fct_reorder_PCB(BMF_species_all_PCB_ng_glw_compare)$PCB)
}
