% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_reorder_species.R
\name{fct_reorder_species}
\alias{fct_reorder_species}
\title{fct_reorder_species Reorder species metadata factors}
\usage{
fct_reorder_species(data)
}
\arguments{
\item{data}{dataset with columns species, labels, grp and alim}
}
\value{
data with reordered factor columns
}
\description{
fct_reorder_species Reorder species metadata factors
}
\examples{
library(chopin.magnification)
levels(fct_reorder_species(contam)$species)
}
