% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{get_stats Calculation of the median, min and max values of a vector from a dataset}
\usage{
get_stats(type, contam, vector)
}
\arguments{
\item{type}{character "Benthos" or "Soles"}

\item{contam}{character ex: name and unit for a chemical concentrations}

\item{vector}{vector ex: chemical concentration for a chemical}
}
\value{
tibble chemical, type, median, min, max
}
\description{
get_stats Calculation of the median, min and max values of a vector from a dataset
}
