% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_censored_with_truncated_fit.R
\name{impute_censored_with_truncated_fit}
\alias{impute_censored_with_truncated_fit}
\title{Imputing values from log-normal distribution for non-detects}
\usage{
impute_censored_with_truncated_fit(data, var, var_cens, distribution = "lnorm")
}
\arguments{
\item{data}{dataset of contamination variables}

\item{var}{character name of the variable containing uncensored data (LOQ values)}

\item{var_cens}{character name of the variable containing censoring information (T/F)}

\item{distribution}{character model distribution (default: "lnorm")}
}
\value{
vector containing imputed values when data was <LOQ
}
\description{
Imputing values from log-normal distribution for non-detects
}
\examples{
impute_censored_with_truncated_fit(data = contam_PFAS_ng_gdw, 
                                   var = "PFNA", var_cens = "PFNA_cen")
}
