% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{max_plus}
\alias{max_plus}
\title{max_plus Calculation of the rounded maximum excluding missing values}
\usage{
max_plus(vector)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
maximal value
}
\description{
max_plus Calculation of the rounded maximum excluding missing values
}
