% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{mean_plus}
\alias{mean_plus}
\title{mean_plus Calculation of the rounded mean excluding missing values}
\usage{
mean_plus(vector)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
mean value
}
\description{
mean_plus Calculation of the rounded mean excluding missing values
}
