% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{median_plus}
\alias{median_plus}
\title{median_plus Calculation of the rounded median excluding missing values}
\usage{
median_plus(x)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
median value
}
\description{
median_plus Calculation of the rounded median excluding missing values
}
