% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{min_plus}
\alias{min_plus}
\title{min_plus Calculation of the rounded minimum excluding missing values}
\usage{
min_plus(x)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
minimal value
}
\description{
min_plus Calculation of the rounded minimum excluding missing values
}
