% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{stats_mean}
\alias{stats_mean}
\title{stats_mean Calculation of the mean, min and max values of a vector}
\usage{
stats_mean(x)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
character \code{mean[min-max]}
}
\description{
stats_mean Calculation of the mean, min and max values of a vector
}
