% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_plus.R
\name{stats_median}
\alias{stats_median}
\title{stats_median Calculation of the median, min and max values of a vector}
\usage{
stats_median(vector)
}
\arguments{
\item{vector}{Vector containing values to be summarized}
}
\value{
character \code{median[min-max]}
}
\description{
stats_median Calculation of the median, min and max values of a vector
}
